+++ src/app/dataset-view/additional-components/metadata-component/tabs/metadata-schema-tab/metadata-schema-tab.component.spec.ts
@@ -16,6 +16,7 @@ import {
 } from "../../../data-tabs.mock";
 import { registerMatSvgIcons } from "src/app/common/helpers/base-test.helpers.spec";
 import { HttpClientTestingModule } from "@angular/common/http/testing";
+import { provideToastr } from "ngx-toastr";
 
 describe("MetadataSchemaTabComponent", () => {
     let component: MetadataSchemaTabComponent;
@@ -24,6 +25,7 @@ describe("MetadataSchemaTabComponent", () => {
     beforeEach(() => {
         TestBed.configureTestingModule({
             imports: [MetadataSchemaTabComponent, HttpClientTestingModule],
+            providers: [provideToastr()],
         });
         fixture = TestBed.createComponent(MetadataSchemaTabComponent);
         component = fixture.componentInstance;